requirements.txt
faker
kafka-python
openai>=1.0.0
python-dotenv
